refactor: atualiza versão das dependências.
