+++ src/App/index.tsx
@@ -1,6 +1,5 @@
 import { FC } from "react";
 
-import { Container } from "./style";
 import Header from "../components/Header";
 import Router from "../router";
 import Footer from "../components/Footer";
@@ -8,12 +7,12 @@ import QuizInLoading from "../components/QuizInLoading";
 
 const App: FC = () => {
   return (
-    <Container>
+    <>
       <QuizInLoading />
       <Header />
       <Router />
       <Footer />
-    </Container>
+    </>
   );
 };
