+++ src/components/QuestionsNumberCard/style.ts
@@ -1,4 +1,5 @@
-import { styled, Box } from "@mui/material";
+import styled from "@mui/material/styles/styled";
+import Box from "@mui/material/Box";
 
 export const Container = styled(Box)({
   gridArea: "action-content"
