+++ src/components/QuizInButton/style.ts
@@ -1,4 +1,5 @@
-import { Button, styled } from "@mui/material";
+import Button from "@mui/material/Button";
+import styled from "@mui/material/styles/styled";
 
 export interface StyleProps {
   position?: "left" | "right";
