+++ src/components/QuizInModal/style.ts
@@ -1,4 +1,5 @@
-import { styled, Modal } from "@mui/material";
+import styled from "@mui/material/styles/styled";
+import Modal from "@mui/material/Modal";
 
 export const Container = styled(Modal)(({ theme }) => ({
   backdropFilter: "blur(10px)",
