+++ src/components/QuizInTableCell/style.ts
@@ -1,4 +1,5 @@
-import { styled, TableCell } from "@mui/material";
+import styled from "@mui/material/styles/styled";
+import TableCell from "@mui/material/TableCell";
 
 export const StyledTableCell = styled(TableCell)({
   border: "none"
