+++ src/types/index.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from "react"
+
+export type WithChildren<T = {}> = {
+  children: ReactNode;
+}  & T;
